+++ packages/devtools_shared/CHANGELOG.md
@@ -1,7 +1,5 @@
-# 8.0.0
+# 7.0.1-wip
 * Refactor yaml extension methods.
-* **Breaking change:** rename `ServerApi.getCompleted` to `ServerApi.success` and make the
-`value` parameter optional.
 
 # 7.0.0
 * **Breaking change:** remove the `ServerApi.setCompleted` method that was a